move colletion to  entity
