+++ Dia/entity/models.py
@@ -144,3 +144,10 @@ class Entity(models.Model):
         except:
             return False
         return True
+
+
+class Collection(models.Model):
+    user = models.ForeignKey('user.User', related_name='related_collection', on_delete=models.CASCADE)
+    ent = models.ForeignKey('entity.Entity', related_name='ent', on_delete=models.CASCADE)
+    type = models.CharField(blank=True, verbose_name='类型', max_length=20)
+    dt = models.DateTimeField(default=datetime.now, verbose_name='文件收藏时间', auto_now_add=True)
