+++ Dia/user/models.py
@@ -1,7 +1,6 @@
 import os
 from datetime import date, datetime, timedelta
 from django.db import models
-
 from user.hypers import *
 
 
@@ -22,14 +21,6 @@ class User(models.Model):
     wrong_count = models.IntegerField(blank=True, verbose_name='最近一天密码错误次数', default=0)
     create_time = models.DateTimeField(blank=True, verbose_name='创建时间', auto_now_add=True)
     profile_photo = models.FileField(blank=True, upload_to=DEFAULT_PROFILE_ROOT, verbose_name="头像路径", max_length=256, default='')
-
-    def verify_vip(self) -> bool:
-        if self.vip_date < date.today():
-            self.identity = 'user'
-            self.save()
-        else:
-            self.identity = 'vip'
-        return self.identity == 'vip'
 
 
 class EmailRecord(models.Model):
@@ -54,8 +45,3 @@ class Message(models.Model):
     type = models.CharField(max_length=20, blank=False, verbose_name='消息类型')
 
 
-class Collection(models.Model):
-    user = models.ForeignKey('user.User', related_name='related_collection', on_delete=models.CASCADE)
-    ent = models.ForeignKey('entity.Entity', related_name='ent', on_delete=models.CASCADE)
-    type = models.CharField(blank=True, verbose_name='类型', max_length=20)
-    dt = models.DateTimeField(default=datetime.now, verbose_name='文件收藏时间')
